+++ Editor/VisualScripting.Core/Analytics/MigrationAnalytics.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
+using UnityEngine.Analytics;
 
 namespace Unity.VisualScripting.Analytics
 {
@@ -11,8 +12,19 @@ namespace Unity.VisualScripting.Analytics
         private const int MaxNumberOfElements = 1000;
         private const string VendorKey = "unity.visualscripting";
         private const string EventName = "VScriptMigration";
+#if !UNITY_2023_2_OR_NEWER
         private static bool _isRegistered = false;
+#endif
 
+#if UNITY_2023_2_OR_NEWER
+        internal static void Send(Data data)
+        {
+            if (!EditorAnalytics.enabled)
+                return;
+
+            EditorAnalytics.SendAnalytic(new MigrationAnalytic(data));
+        }
+#else
         internal static void Send(Data data)
         {
             if (!EditorAnalytics.enabled)
@@ -37,9 +49,32 @@ namespace Unity.VisualScripting.Analytics
 
             return _isRegistered;
         }
+#endif
+
+#if UNITY_2023_2_OR_NEWER
+        [AnalyticInfo(eventName: EventName, vendorKey: VendorKey, maxEventsPerHour:MaxEventsPerHour, maxNumberOfElements:MaxNumberOfElements)]
+        class MigrationAnalytic : IAnalytic
+        {
+            private Data data;
+            public MigrationAnalytic(Data data)
+            {
+                this.data = data;
+            }
+
+            public bool TryGatherData(out IAnalytic.IData data, out Exception error)
+            {
+                error = null;
+                data = this.data;
+                return data != null;
+            }
+        }
 
         [Serializable]
+        internal class Data : IAnalytic.IData
+#else
+        [Serializable]
         internal class Data
+#endif
         {
             [SerializeField]
             internal MigrationStepAnalyticsData total;