+++ Editor/VisualScripting.Flow/Analytics/FlowMacroSavedEvent.cs
@@ -1,20 +1,10 @@
-using UnityEditor;
-
 namespace Unity.VisualScripting.Analytics
 {
     class FlowMacroSavedEvent : UnityEditor.AssetModificationProcessor
     {
         static string[] OnWillSaveAssets(string[] paths)
         {
-            foreach (string path in paths)
-            {
-                var assetType = AssetDatabase.GetMainAssetTypeAtPath(path);
-                if (assetType == typeof(ScriptGraphAsset))
-                {
-                    UsageAnalytics.CollectAndSend();
-                    break;
-                }
-            }
+            UsageAnalytics.CollectAndSendForType<ScriptGraphAsset>(paths);
             return paths;
         }
     }