+++ Runtime/VisualScripting.Flow/Framework/Codebase/MemberUnit.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
 using UnityObject = UnityEngine.Object;
 
 namespace Unity.VisualScripting